requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.1.0b1
psycopg[binary]==3.3.3
alembic>=1.13.1
python-dotenv==1.0.0
PyJWT==2.8.1
eth-account==0.10.0
eth-keys==0.6.0
eth-utils==2.1.1
cryptography==41.0.7
pytest==7.4.3
httpx==0.25.2
ipfshttpclient==0.8.0
pydantic==2.5.0
python-multipart==0.0.6

